backend/requirements.txt
# FastAPI
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.11.0
pdf2image==1.17.0

# Image Processing
Pillow==10.2.0
opencv-python-headless==4.9.0.80
numpy==1.26.4

# Export
reportlab==4.1.0
fpdf2==2.7.9

# AI Integration
httpx==0.27.0
google-auth==2.28.0

# Utils
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
aiofiles==23.2.1
